refactor: FollowUSerData Model refactoring
